requirements.txt
requests
google-api-python-client
python-dotenv
psycopg2-binary
aiogram
